requirements.txt
TensorFlow==2.13.0
streamlit==1.25.0
streamlit-authenticator==0.2.2
numpy
seaborn
nltk==3.8.1
assemblyai==0.15.1
transformers==4.22.1
scipy==1.8.0
detoxify==0.5.1
pyotp==2.9.0
qrcode==7.4.2
cryptography==39.0.2
streamlit_extras
sqlalchemy
azure-identity
azure-keyvault-secrets
